requirements.txt
pyTelegramBotAPI
python-dotenv
aiohttp
fastapi
uvicorn